Added Entity class and changed platform collision
Added a abstract class called Entity. Will make it easier handle collision. Player and enemies (if we want some) will derive from the entity class.

Added a Enemy Class (unfinished)

Changed HandleCollision() inside of PlatformHandler to take in an entity and check collision with it instead of just only the player.

Moved some of the players variables and methods to the Entity class for example GetPosition(), since it could be nice to let other objects that derive from entity t use them aswell.

Since I moved a few variables I also changed name on them to be generic, ex PlayerSprite is now sprite. So I had to change a few classes to fit this change.